handle all book operations while persisting book data on local storage
